Moved some dom creation tasks to seperate modules.
Also updated edting task to now select the project already set for
that task
